🔍 Refactor `/user` dto
