+++ src/user/dto/create-user.dto.ts
@@ -1,6 +1,4 @@
 import { ApiProperty } from '@nestjs/swagger';
-import { HttpResponseDto } from '../../HttpResponseDto';
-import { User } from '../entities/user.interface';
 
 export class CreateUserDto {
   @ApiProperty({
@@ -53,58 +51,4 @@ export class CreateUserDto {
   //   default: ['review1', 'review2'],
   // })
   // review_list: string[];
-}
-
-export class Success_CreateUserDto extends HttpResponseDto {
-  @ApiProperty({
-    default: 201,
-  })
-  statusCode: number;
-
-  @ApiProperty({
-    example: '유저 생성이 성공적으로 처리되었습니다.',
-  })
-  message: string;
-
-  @ApiProperty({
-    default: false,
-  })
-  error: boolean;
-
-  @ApiProperty({
-    default: {
-      id: 'test-id',
-      name: 'test-name',
-      bio: 'test-bio',
-      profileImageURL: 'test-profileImage-url',
-      role: 'student',
-    },
-  })
-  data: User;
-
-  constructor(data: User) {
-    super(201, '사용자 생성이 성공적으로 처리되었습니다.', false, data);
-  }
-}
-
-export class BadRequest_CreateUserDto extends HttpResponseDto {
-  @ApiProperty({
-    example: 'Bad request.',
-  })
-  message: string;
-
-  @ApiProperty({
-    default: true,
-  })
-  error: boolean;
-
-  @ApiProperty({
-    default: 400,
-  })
-  statusCode: number;
-
-  @ApiProperty({
-    default: null,
-  })
-  data: User;
 }
